+++ irs-registry-client/src/test/java/org/eclipse/tractusx/irs/registryclient/decentral/DecentralDigitalTwinRegistryServiceTest.java
@@ -203,11 +203,12 @@ class DecentralDigitalTwinRegistryServiceTest {
                 expectedShell);
 
         // when
-        final Collection<String> globalAssetIds = decentralDigitalTwinRegistryService.lookupGlobalAssetIds(
+        final Collection<AssetAdministrationShellDescriptor> assetAdministrationShellDescriptors = decentralDigitalTwinRegistryService.lookupShellsByBPN(
                 digitalTwinRegistryKey.bpn());
 
+        String actualGlobalAssetId = assetAdministrationShellDescriptors.stream().findFirst().map(AssetAdministrationShellDescriptor::getGlobalAssetId).get();
         // then
-        Assertions.assertThat(globalAssetIds).containsExactly(expectedGlobalAssetId);
+        Assertions.assertThat(actualGlobalAssetId).isEqualTo(expectedGlobalAssetId);
     }
 
 }